Measure a fold the provider was too late to report, and close two classes by sweep
A provider session that says nothing when it is adopted no longer keeps a fold of
zero for the rest of its life. IContextUsageReporter permits CurrentUsage to be null
at creation and says an implementation that sometimes knows returns null until it
does, so the one instant the fold was measured is the one instant such a provider is
silent. The measurement is deferred rather than settled: it is taken at the first turn
the provider does report on, against this library's estimate of the conversation by
then, and is recorded as unmeasured in the meantime so a genuine zero is not
re-measured against a conversation that has since grown. Refusing the reporting
transition was the alternative and was rejected - it abandons a session over
conformant adapter behavior, after a real message has been spent.

Integer overflow, swept rather than patched. The transcript's fit test compares by
subtraction and its cached total is accumulated wide and refused where it exceeds a
token count; the layout's conversation and total figures saturate instead of wrapping;
the in-memory session's overhead and usage sums do the same; the tool declaration
terms are widened before they are added. Every other combination in the package is
bounded by the string cap recorded on TokenEstimator.EstimateTokens.

Whitespace-as-empty, swept to the consumer that charged it. A tier's cached estimate
is zero for a blank record, so IsWithinBudget and every other consumer that measures a
tier reach the same answer as IsEmpty, ConversationTokens and BuildSeed. The rotation
engine's saturation input and cascade decision read the tier rather than re-estimating
its text, which stops a blank record inflating the input a consolidation is measured
against and suppressing a redundancy signal it had earned.

The reviewer's overflow in (text.Length + 3) / 4 is refuted rather than implemented: a
string is one object, the runtime caps one object at two gigabytes, and the largest
allocatable length measured on these targets is 1,073,741,791. The cap is recorded in
the code so the next review reaches the same conclusion without re-deriving it.

The forced-path rotation test stops trading on a pairing claim it cannot make. At a
budget of zero the snap decides nothing, so the test is renamed and re-documented as
what it does establish - the whole history, in recorded order - and the reqstream trace
follows.

+++ src/DemaConsulting.AgentKit.Sessions/TokenEstimator.cs
@@ -77,10 +77,27 @@ public static class TokenEstimator
     ///     Estimates the tokens a run of text occupies in a context window.
     /// </summary>
     /// <remarks>
+    ///     <para>
     ///     Rounds up, so a non-empty string never estimates as zero tokens: a budget comparison
     ///     that treated short content as free would let an unbounded number of short entries
     ///     accumulate. Null and empty text estimate as zero, which is the honest answer for
     ///     absent content.
+    ///     </para>
+    ///     <para>
+    ///     <b>The rounding addition cannot overflow, and this is the invariant that says so.</b>
+    ///     A <see cref="string"/> is a single object, and the runtime caps one object at two
+    ///     gigabytes — <c>gcAllowVeryLargeObjects</c> raises that for arrays and not for strings —
+    ///     so at two bytes per character the longest string that can exist holds a little under
+    ///     2^30 characters. Measured on this repository's own targets, the largest
+    ///     <c>new string('a', n)</c> that allocates is n = 1,073,741,791 and n = 1,073,741,792
+    ///     throws <see cref="OutOfMemoryException"/>. <c>Length + 3</c> therefore reaches at most
+    ///     1,073,741,794, which is short of half of <see cref="int.MaxValue"/>, and the quotient
+    ///     reaches at most 268,435,448 tokens. This is recorded here because it is the bound every
+    ///     other estimate in this package inherits: a single estimate can never be negative, and
+    ///     two or three of them can be added in plain token arithmetic without wrapping. It has
+    ///     been raised as an overflow twice; the arithmetic is correct and the reason is written
+    ///     down rather than restated in a review each time.
+    ///     </para>
     /// </remarks>
     /// <param name="text">
     ///     The text to estimate. May be <see langword="null"/> or empty, both of which estimate as
@@ -97,6 +114,10 @@ public static class TokenEstimator
         }
 
         // Round up so short-but-present content is charged at least one token.
+        //
+        // The addition is in plain token arithmetic and stays there: a string cannot hold two
+        // gigabytes, so Length is capped a little under 2^30 and Length + 3 cannot approach
+        // int.MaxValue. See the remarks above for the measurement behind that cap.
         return (text.Length + CharactersPerToken - 1) / CharactersPerToken;
     }
 
@@ -157,12 +178,21 @@ public static class TokenEstimator
         // structure wrapped around them.
         //
         // Accumulated in a wider type than a token count, and range-tested before it is narrowed.
-        // Each declaration fits an int on its own, because a string cannot be longer than
-        // int.MaxValue characters and the ratio only divides; their sum need not, and an int
+        // Each declaration fits an int on its own, because a string cannot be longer than the
+        // runtime's object cap allows and the ratio only divides; their sum need not, and an int
         // accumulator would wrap it to a small or negative figure that every caller downstream
         // would then treat as a real measurement of the overhead. The declarations are the point at
         // which this figure first becomes computable, so it is rejected here and no later site has
         // to ask again.
+        //
+        // Each addend is widened before any of them are added, rather than after. Written as
+        // 'total += a + b + c + d' the four int estimates are summed in int arithmetic and only the
+        // result is widened, which makes the wide accumulator depend on a second invariant to be
+        // sound: that one declaration's three estimates cannot themselves wrap. They cannot - see
+        // EstimateTokens, where the string cap holds every estimate below 2^28 and three of them
+        // below 2^30 - so the previous form was not in fact defective. It is widened here anyway
+        // because an accumulator that exists to catch an overflow should not be reached through
+        // arithmetic that could have one, and because the cost is nothing.
         long total = 0;
         foreach (var tool in tools)
         {
@@ -171,7 +201,7 @@ public static class TokenEstimator
                 throw new ArgumentException("A tool in the list is null.", nameof(tools));
             }
 
-            total += EstimateTokens(tool.Name)
+            total += (long)EstimateTokens(tool.Name)
                 + EstimateTokens(tool.Description)
                 + EstimateTokens(tool.JsonSchema.ToString())
                 + PerToolOverheadTokens;
